requirements.txt
Pillow==9.5.0
requests==2.31.0
numpy==1.24.3
torch==2.0.1 # veya tensorflow==2.13.0
torchvision==0.15.2
opencv-python==4.8.0.74
transformers==4.35.0
beautifulsoup4==4.12.2
spacy==3.6.1
nltk==3.8.1
google-api-python-client==2.97.0 # Google arama için
pyyaml==6.0.0
tqdm==4.66.1 # İlerleme çubuğu için